refactor: change test file location
