+++ src/components/layout/Layout.test.tsx
@@ -1,0 +1,16 @@
+import { render, screen } from '@testing-library/react';
+import { describe, it } from 'vitest';
+import '@testing-library/jest-dom';
+import { MemoryRouter } from 'react-router-dom';
+import { Layout } from './Layout';
+
+describe('Home', () => {
+  it('renders main', () => {
+    render(
+      <MemoryRouter>
+        <Layout title="Home" content={'Home'} />
+      </MemoryRouter>
+    );
+    expect(screen.getByRole('main')).toBeInTheDocument();
+  });
+});
