+++ src/pages/home/Home.test.tsx
@@ -1,0 +1,18 @@
+import { render, screen } from '@testing-library/react';
+import { describe, it } from 'vitest';
+import '@testing-library/jest-dom';
+import { MemoryRouter, Route, Routes } from 'react-router-dom';
+import { Home } from './Home';
+
+describe('Home', () => {
+  it('renders Home route', () => {
+    render(
+      <MemoryRouter initialEntries={['/']}>
+        <Routes>
+          <Route path="/" element={<Home />} />
+        </Routes>
+      </MemoryRouter>
+    );
+    expect(screen.getByText(/home/i)).toBeInTheDocument();
+  });
+});
